analytics/cloud_functions/extract_resources/requirements.txt
google-cloud-storage>=2.12.0
